Assistant checkpoint: Convert NumPy types to native Python types for database
Assistant generated file changes:
- utils/simulator.py: Convert NumPy types to native Python types

---

User prompt:

Error during simulation: (psycopg2.errors.InvalidSchemaName) schema "np" does not exist LINE 1: ... '2025-02-10T00:00:00+00:00'::timestamptz, 10000, np.float64... ^

[SQL: INSERT INTO simulation_results (strategy_id, symbol, start_date, end_date, initial_balance, final_balance, total_trades, win_rate, sharpe_ratio, max_drawdown, created_at) VALUES (%(strategy_id)s, %(symbol)s, %(start_date)s, %(end_date)s, %(initial_balance)s, %(final_balance)s, %(total_trades)s, %(win_rate)s, %(sharpe_ratio)s, %(max_drawdown)s, %(created_at)s) RETURNING simulation_results.id] [parameters: {'strategy_id': 1, 'symbol': 'BTC-USD', 'start_date': Timestamp('2024-02-10 00:00:00+0000', tz='UTC'), 'end_date': Timestamp('2025-02-10 00:00:00+0000', tz='UTC'), 'initial_balance': 10000, 'final_balance': np.float64(160783.20440514054), 'total_trades': 0, 'win_rate': np.float64(0.3469945355191257), 'sharpe_ratio': np.float64(7.425288776982815), 'max_drawdown': np.float64(0.9378044489348398), 'created_at': datetime.datetime(2025, 2, 10, 14, 28, 39, 299080)}] (Background on this error at: https://sqlalche.me/e/20/f405)

+++ utils/simulator.py
@@ -74,12 +74,12 @@ class TradingSimulator:
             symbol=symbol,
             start_date=start_date,
             end_date=end_date,
-            initial_balance=self.initial_balance,
-            final_balance=self.initial_balance * (1 + metrics['Total Return']),
+            initial_balance=float(self.initial_balance),
+            final_balance=float(self.initial_balance * (1 + metrics['Total Return'])),
             total_trades=0,  # To be implemented
-            win_rate=metrics['Win Rate'],
-            sharpe_ratio=metrics['Sharpe Ratio'],
-            max_drawdown=metrics['Max Drawdown']
+            win_rate=float(metrics['Win Rate']),
+            sharpe_ratio=float(metrics['Sharpe Ratio']),
+            max_drawdown=float(metrics['Max Drawdown'])
         )
 
         db.add(simulation)
